+++ tests/testthat/_snaps/renv_helpers/default_v0-14-0.nix
@@ -190,7 +190,9 @@ let
     inherit (pkgs) 
       R
       glibcLocales
-      nix;
+      nix
+      pandoc
+      which;
   };
   
   shell = pkgs.mkShell {