+++ tests/testthat/_snaps/renv_helpers/default_v0-15-5.nix
@@ -175,7 +175,9 @@ let
     inherit (pkgs) 
       R
       glibcLocales
-      nix;
+      nix
+      pandoc
+      which;
   };
   
   shell = pkgs.mkShell {