+++ tests/testthat/_snaps/renv_helpers/default_v0-17-3.nix
@@ -156,7 +156,9 @@ let
     inherit (pkgs) 
       R
       glibcLocales
-      nix;
+      nix
+      pandoc
+      which;
   };
   
   shell = pkgs.mkShell {