+++ tests/testthat/_snaps/renv_helpers/default_v1-0-7.nix
@@ -83,7 +83,9 @@ let
     inherit (pkgs) 
       R
       glibcLocales
-      nix;
+      nix
+      pandoc
+      which;
   };
   
   shell = pkgs.mkShell {